Added to the strings.xml, drawable folder, and RexDatabaseHelper.java
Added all the symptoms from the allergies to the strings.xml
- I figured this way it reduce the need to repeat code/typing for
foods that have the same symptom.
- most symptoms are labeled numerically, some need to be specified by
the food i.e. symptom_onion_family

Added the insert methods for FOOD, DOG, and ALLERGIES databases
- changed the QUOTE STRING to QUOTE INTEGER in the FOOD table
so that it can take the string for the symptoms of the food
from string.xml

Added all the foods with pictures and symptoms to FOOD database
-used the insertFood method to populate the FOOD database

Added all pictures to the drawable folder
- all the pictures are in png format

Signed-off-by: Gavin Erezuma <gerezuma@mills.edu>