Added to RexDatabaseHelper.java
This did not commit the first time

Added all the symptoms from the allergies to the strings.xml
- I figured this way it reduce the need to repeat code/typing for
foods that have the same symptom.
- most symptoms are labeled numerically, some need to be specified by
the food i.e. symptom_onion_family

Added the insert methods for FOOD, DOG, and ALLERGIES databases
- changed the QUOTE STRING to QUOTE INTEGER in the FOOD table
so that it can take the string for the symptoms of the food
from string.xml

Signed-off-by: Gavin Erezuma <gerezuma@mills.edu>